added a popup for food deletion
the popup displays the meals that will be deleted as a result of the food being deleted and allows the user to cancel or confirm the deletion. It is also able to change the percentages to add up to 100. I added the meal object to the loadFoods method in scene controller in order to display the foods in foodForm